refactor: move migrationMonitor out of live-migration-source.go
live-migration-source.go had grown into a large file mixing the
migrationMonitor implementation with unrelated migration logic.

Move the migrationMonitor struct, its constants, downtimeTuningConfig,
newMigrationMonitor/newDowntimeTuningConfig and all migrationMonitor
methods into a dedicated migration-monitor.go, and the corresponding
specs into migration-monitor_test.go. No functional change.

Fixes #18184
