+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "live-migration-source.go",
         "live-migration-target.go",
         "manager.go",
+        "migration-monitor.go",
         "stall-detector.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap",
@@ -102,6 +103,7 @@ go_test(
         "live-migration-source_test.go",
         "live-migration-target_test.go",
         "manager_test.go",
+        "migration-monitor_test.go",
         "virtwrap_suite_test.go",
     ],
     data = glob(["testdata/**"]),
